Add files via upload
Well this took a lot of tries, definitely more than 30 and 3 times I had to start from the start because references can fix your code but they surely will mess it up if they don't fix it(at least for me).
Scored 96/100, was stuck at around 69/100 then after a lot of time of staring at the code found out a single -1 not existing in the conversion from 2d index to 1d index was ruining my time and life.
Fixed it after all and managed to pass. 11th November 2023
Hare Krishna.